App::perlbrew: use "test_harness" on perl >=5.7.3
Change the default perlbrew behavior so that it now tests via "make
test_harness" instead of "make test" if the "test_harness" target is available.

With the "test_harness" target we'll get automatic parallel testing
via $HARNESS_OPTIONS.

The + "test_harness" target was added in 5.7.3, which was the last
development release before 5.8.0.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>