Don't symlink executables into bin/
If we do, "perlbrew use" doesn't work, because when it overrides your path
to "bin/:perls/VERSION/bin", the symlinks in bin/ will take precedence.

bin/ needs to be kept clean of "perl" commands for use to work.

There is one case left: "perlbrew switch /path/to/very/specific/perl"
"perlbrew use" still fails to work with that one.

We also included upgrade notes in the Changes file. We need to

   perlbrew init
   perlbrew off
   perlbrew switch version-you-want

to clean up all the old symblinks.

Signed-off-by: Pedro Melo <melo@simplicidade.org>