Make sure we export the SHELL environment
On some sequences of commands, the perlbrew perl command didn't get the
SHELL environment. This was noticed inside Jenkins CI jobs, for example.

We suspect that the SHELL variable was set by Jenkins but not exported,
and so external commands don't see it.

But we make some decisions based on it, most notably the perlbrew env
command.

This export SHELL will make sure we always get it.

Signed-off-by: Pedro Melo <melo@simplicidade.org>