run 'symlink-executables' automatically after installing a dist without an executable named 'perl'.
This is majorly for dev versions of perl. For example, perl-5.13.11
install perl executable under the name 'perl5.13.11' but not 'perl'.

In fact, all execuatbles installed by perl-5.13.11 are suffixed with
5.13.11.  The purpose of 'symlink-executables' is to simlink those
executables to a name without version suffix, in order to make
'perlbrew switch' or 'perlbrew use' commands work smoothly.