add support for multiple .perlbrew dirs
If your home directory is shared across multiple systems you may want
to have a different perlbrew installed for each system.  Perlbrew is
already largely set up to do this, but unfortunately the
.perlbrew/init file path has been hardcoded into perlbrew.  This patch
looks for a new environment variable, $PERLBREW_HOME, to decide where
to store init.  If $PERLBREW_HOME is not found, it defaults to
$HOME/.perlbrew.