Refactor: let installed_perls returns a array of hashref. And improve 'perllbrew list' command.
Each element in the array should contain key-value paris for 3 keys: 'name', 'version', 'is_current',
that looks like:

     {
         name => "perl-5.12.3",
         version => "5.12.3",
         is_current => 1
     }

name is usually 'perl-<version>', but people may renamed the installation with '--as' option.
That's why it's better to also provide a 'version' information.

`perlbrew list` command is improved in the way that, if the installation name does not
contain it's own version number, then show the version number afterwards.

For example:

    > perlbrew list
      perl-5.12.2
    * perl-5.12.3
      perl-5.13.10
      perl-5.14.0-RC2
      perl-blead (5.14.0)
      perl-v5.14.0-RC1-8-g2cf7ccf
      /usr/bin/perl (5.10.0)

It shows version number for perl-blead and /usr/bin/perl