re-write the 'help' command to pull output from pod.
So the idea is, `perldoc perlbrew` should produce the full user manual,
including *everything* user should know to use perlbrew.

`perlbrew help` should produce a full list of help topics, so user
know that they can run `perlbrew help <topic>`

`perlbrew help <topic>` should produce the documentation for the topic.

In most cases, a topic is a command name. User should be able to run `perlbrew
help install` or `perlbrew help lib` to see the full usage of that command.

When a topic is not a command, it should be made clear so user don't get
confused.