Add sitecustomize option
This commit adds the --sitecustomize option to allow users to specify a
sitecustomize.pl to be installed when Perlbrew builds a perl. This is
helpful because sitecustomize.pl can be used (among other things) for
dynamic modifications to @INC.  The option sets -Dusesitecustomize and
copies the specified file to the newly built $Config{sitelib}.

The commit also adds Capture::Tiny as a prerequisite to allow for
capturing via do_system() instead of using qx//, which allows sufficient
mockability to test the behavior of sitecustomize without actually
building a Perl.