require the core Cwd early to avoid arch conflict in PERL5LIB
The issue is discovered @dagolden in GH #297

Basically we need to ensure the core version of Cwd is loaded
instead of its twin brother in PERL5LIB, which might be built
with different archname.

While using lib::core::only works with fatpacked executables,
it breaks CPAN installation. This change with several others
beforhand should work for both scenario. However at some point
it might be more convienent to release the fatpacked execuatable
to CPAN too.