briefly check if the downloaded cpanm or perlbrew seems to be a valid perl program, and die with some reasonable message if not.
Since we download them from github and github has been in maintainence
frequently enough to make this a sensible issue, this commit perform a super
dummy check on the downloaded content to see if they begin with "#!/" and
decided whether they are valid or not.

Of course this is far from accurate, but let's see if it is helpful.