Add support for a ".version" file in each perls root directory.
This is somewhat better than executing each and every perl on your
system for each and every new shell session.

Though this patch only adds support for such a file if it exists,
it does not yet cause such files to exist, and it only serves so I can
manually fix perlbrew from spewing in every terminal profusely.

[rt#92942](https://rt.cpan.org/Ticket/Display.html?id=92942)