Fix to MANPATH
This is a fix for fish so that the MANPATH will be a
':' delimited string instead of an array (expands to
words separated by spaces).  If this is not colon-delimited,
then the /etc/manpath.config is not sourced, (and) man
may not use all of the paths in MANPATH when looking for files.