redo daa2c94a370f959b9616b4b7fca0b2ef2cad99a2 for bash and zsh
Since I generally want minimum differences for our bash and zsh
scripts, we need to have a branch that depends on $BASH_VERSION
or $ZSH_VERSION to tell us what shell we are in. Because unfornatedly,
the "read" builtin has different parameter for reading parameters
into an array ( "-a" for bash, "-A" for zsh ).