prefer https and remove insecure download cli flags.
This has been flipped back-and-forth several times in the path, and we
tended to favor convienence over security issues. Which is less ideal
and probably should be changed.

With this commit, we are change several parts to make the downloads done
internally a bit safer -- at least for "patchperl" and "cpanm" programs.

If in future we see issues about certificate verification failures, we
should "fix" them by providing instructions to setup curl/wget/fetch
properly.