A copy 64161eb30fdc349ca338a36cb695c57a221a6a9b in fatpack version.
The purpose of these commits is to nullify the effect of PERL5LIB when
doing `require Cwd` -- so that we always load the Cwd shipped with
distribution. Otherwise perlbrew itself might be loading a Cwd that
is binary-incompatible with system perl. Since perlbrew itself is
supposed to be executed with system perl, that basically makes it
fail to be compiled.