Introduced method available_perls_with_urls.
This method provides a way to print out available perl installations with URLs from which they are going to be download. This can help users to know where a perl version will be installed.
To provide such functionality the method will return an hash reference with the perl version and the url as value. The method available_perls has been refactored to extract only the keys of such hash.