The available command has been extends in order to print URLs along with Perl version.
The idea is that the user should know from where each version is downloadable.
In order to achieve this, the new method available_perls_with_urls has been added,
it returns an hash reference where the key is the perl version and the value is the
url for such version.

The method available_perls have been refactored to return the keys of available_perls_with_urls,
so that the parsing mechanism happens once.

The test has been adapted to work with the new implementation.