Save list-modules output to a temporary file.
The run_command_list_modules now accepts an optional argument as a filename.
If such filename is provided, the sub-perl process opens the file
and prints to such file instead of to direct STDOUT.
This can be helpful for later processing of the command output.

The run_command_clone_modules now uses File::Temp to pass a temporary
filename to which the program must write the list of modules in order
to get the list back at the time of installation into another Perl instance.