Sort the available Perl version from the newest to the oldest.
In the 'available' command it is annoying to have the available Perl list
with the oldest versions on top of the screen and the last on the bottom.
Users are going to install newer versions most of the time, so their first
choices should be the newest version, and therefore they should be available
on the top of list.
The problem is oldest version did not follow a coherent version numbering,
and therefore it is required to mangle the version number in order to make
it comparable.

In this implementation I transform a Perl version into a comparable string,
so that for instance v5.24.1 becomes 50_024_001, and so it can be
sorted quite easily with a Schwartzian-transform.

In order to keep 'Perl' separated from 'cperl', the latter version number
is translated into a string with no 10 factor multiplier on the
major version number.