Add the ctime information to the hash of installed perls.
This allows for a quick reminder of when a Perl installation has been performed.
The idea is to stat(1) the executable of an installed Perl instance
to get the 'ctime', that is when the executable has been created.
Such timestamp is evaluated via localtime in scalar context to get a string
human-readable of the timestamp.

The list command now presents the information about installation time.