Use the sorting of Perl versions across the other commands.
Created a method to sort the perl versions so that it is easier
to refactor other subcommand to present the perl list ordered
in the same way.

Refactor the available_perls method to use the object oriented interface.
Refactor the run_command_available method to return the sorted list
of perl instances.