Add the same sorting from newest to oldest in the 'list' command.
In order to achieve the ordering, the hash of installed perls have now
a different key named 'comparable_version' that is the result of
the comparable_perl_version method. The ordering of the @result array
is now performed on such version now and not on the orig_version.
I left the name comparison untouched since I'm not sure it has to be
inverted ($b vs $a).