Check for .profile instead of specific shell configuation file.
Shoudl fix #563

For bash and zsh it is possible to have user configuration into a
.profile file, so in that case it should not be suggested to place the
configuration for perlbrew into .bash_profile (etc.), but to append
it to the ~/.profile file instead.
This piece of code checks for the $yourshrc and ~/.profile
in the user home. The first found (probably already $yourshrc)
is kept.