Keep options '--force' and '--yes' coherent across commands.
Fix #539 (partially).
There are command that accept the '--force' option and others that
accept the '--yes' command (particularly those that interact with
external modules or utilities). With this simple fix each time either
'yes' or 'force' has been enabled the other is enabled too, so to keep
the code untouched and the options coherent.