concise the type of return value of comparable_perl_version
These types of scalar inputs are (de facto) considered to be valid:

- "cperl-5.16.1"
- "perl-5.16.0"
- "5.8.0"

They can roughly be described as:

    version numbers, optionally prefixed by "cperl-", or "perl-".

There are some tests producing warnings about using non-numerical values
such as "5.6.0" with a "<" operator. This commit fixes those warnings too.

This commit also turns non-valid input into a dummy default value of 0,
Which means.... if there are some perl version strings we do not recognize,
we do not care about how they are sorted.