Add hook to reverse sorting of Perl version.
As described in issue #574, some users do prefer to have
the sorting of Perl versions from the oldest to the newest and not
the opposite as it happens now.

Now the object of the App:Perlbew can contain a variable 'reverse'
that can be parametrized to force any other sorting operation, in particular
the 'sort_perl_versions' subroutine. The idea is to keep the
option for sorting as an object property to keep coherence across
different commands.