take MAKE env var as the 'make' program.
A followed up of 62e3c5b49fb77a5524e55d7c34df38f6dcddfed4.

As discussed in [issue 600][1] for the purpose of using perlbrew on
Solaris systems, "gmake" needs to be used. MAKE env is one approach.

An extra check on osname is also included to use "gmake" by default
on Solaris when MAKE env is missing.

As a result, MAKE env can also be used in case some implementation of
"make" is preffered.

[1]: https://github.com/gugod/App-perlbrew/issues/600