New implementation of the comparable_perl_version function.
Now there are two branches to select a version: a numeric one and a blead one.
The former works pretty much as in the past, converting a version number
to a string made by MMmmpp00 (MM = Major, mm = minor, pp = patch).
The "blead" branch converts it to 99999901 to indicate a major version number.
Since there could be only one blead, it should be ok to keep the number
fixed.