Fix a test failures caused by not preserving file modes in tarballs.
The file modes are lost in the outcome of `mbtiny distdir` and that
breaks the assumption of `t/http-ua-detect-non-curl.t` here.

This commit ensure the file mode of `t/fake-bin/curl` to be an
executable, before performaning the tests.

The alternative solution is of course modifying `mbtiny` to
preserve file modes, but that'll perhaps trigger different
kind of problems.

Perhaps it is just bettter not depending on the current on-disk state
of the module author at the moment of releasing.

Fix issue #615