Fix clone-modules.
See issue #632.
The problem is that the current implementation is always using 'current_oerl'
as the one on which 'run_command_list_modules' is executed, therefore it
does clone only modules from the current perl instance.

This implementation creates a new application with the right version of perl,
that could still be current if none has been resolved, and performs the module
listing.

The run_command_list_modules function has been removed, it seems
to not be called anymore.