add a `list_modules` method.
This method returns an ArrayRef[ModuleName] that is the list of
installed modules for specified env.

This is so the result can be reused for both `list-modules` command as
well as `clone-modules` command

Internally it renames a few selected outputs that are known to be a
little bit "inconveninet" for cpanm.