make-pp: add privlib
privlib usually contains all "core modules".

Core modules are only included in PAR if they are detected
as a dependency by pp. In case of detection failures, the
only way to manually put a core module in a PAR would be
to add -M paramaters.

Here is one such failure case:

    # pp --bundle  -o ack =ack

    # ls | ./ack a
    Undefined subroutine &Term::ANSIColor::colored called at script/ack line 995, <STDIN> line 1.

    # unzip -l ack | grep Term
        13043  08-23-2023 09:37   lib/Term/Cap.pm

About Term::ANSIColor

    # corelist Term::ANSIColor

    Data for 2023-07-02
    Term::ANSIColor was first released with perl v5.6.0

    # perldoc -l Term::ANSIColor
    /home/gugod/perl5/perlbrew/perls/perl-5.38/lib/5.38.0/Term/ANSIColor.pm

The obvious way to overkill this problem is to always include the entire
privlib in the PAR.