work with distdir instead of tarball.
Not sure what part of github workflow has changed, but having a
tarball in artifact (which itself is a zip-archive) would lead us to a
weird situation that the name of the directory instide the tarball
becomes the the as the the tarballs itself, and makes `tar xfz` always
fails.

I guess it can be bypassed by avoid making a tarball in the first place.